tests/test_monitoring.py
```python
# Importing required libraries
import unittest
import time
from unittest.mock import patch, MagicMock
from src import monitoring

class TestMonitoring(unittest.TestCase):
    def setUp(self):
        self.monitor = monitoring.Monitoring()

    @patch('time.sleep', return_value=None)  # Mocking time.sleep function
    def test_process_request(self, mock_sleep):
        """
        Test if the process_request method is working correctly.
        """
        self.monitor.process_request(1)
        mock_sleep.assert_called_once_with(1)

    @patch('datadog.api.Timeboard.create', return_value=None)  # Mocking Datadog API
    def test_send_metrics_to_datadog(self, mock_datadog):
        """
        Test if the send_metrics_to_datadog method is working correctly.
        """
        self.monitor.send_metrics_to_datadog('test_metric', 1)
        mock_datadog.assert_called_once()

    @patch('newrelic.agent.add_custom_parameter', return_value=None)  # Mocking New Relic Agent
    def test_record_custom_metrics_newrelic(self, mock_newrelic):
        """
        Test if the record_custom_metrics_newrelic method is working correctly.
        """
        self.monitor.record_custom_metrics_newrelic('test_metric', 1)
        mock_newrelic.assert_called_once_with('test_metric', 1)

if __name__ == "__main__":
    unittest.main()
```
